refactor: EvaluatorFactory 제거, 함수로 교체
