+++ frontend/src/modules/evaluator/types.ts
@@ -9,10 +9,4 @@ export type EvalResult = {
   result: unknown;
 };
 
-export type Evaluator = {
-  isIdle: boolean;
-  worker: Worker;
-  currentTask: EvalMessage | null;
-};
-
 export type TaskEndMessage = EvalResult & { task: EvalMessage | null };
